backend/SicopBack/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
django-simple-history==3.10.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.7
pillow==11.3.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2

